+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 ## 1.6 branch
 ### 1.6.4
 * `object_map()` global function throws the new
-  `Tools\Exception\MethodNotExistsException`;
+  `\Tools\Exception\MethodNotExistsException`;
 * `\Tools\Exceptionist::methodExists()` throws the new
     `Tools\Exception\MethodNotExistsException`;
+* by default, the `\Tools\Exceptionist::____callStatic()` magic method
+    now throws an `ErrorException`;
 * improved the description of many `Exceptionist` magic methods;
 * `\Tools\TestSuite\TestTrait::assertException()` correctly ignores the 
   deprecations in determining the exception;